// src/app/interfaces/summary.interface.ts
export interface ISummary {
  first_name?: string;        // ✔️
  last_name?: string;         // ✔️
  summary_objective?: string; 
  email?: string;             // ✔️
  phone?: string;             // ✔️
  location?: string;          // ✔️
  url?: string;               // ✔️
  alt_url?: string;           // ✔️
  twitter_handle?: string;    // ✔️
  linkedin_handle?: string;   // ✔️
  github_handle?: string;
}